SCMS FULL/SHOP/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
PyQt5==5.15.9
requests==2.31.0
Pillow==10.2.0 